[mapleall]Internal commit msg:     [mapleall] PCLint warning fix 2.

+++ src/mapleall/maple_be/include/cg/pressure.h
@@ -25,7 +25,7 @@ struct RegList {
 };
 
 #define FOR_ALL_REGCLASS(i) \
-  for (int32 i = 0; i < RegPressure::GetMaxRegClassNum(); ++i)
+  for (uint32 i = 0; i < static_cast<uint32>(RegPressure::GetMaxRegClassNum()); ++i)
 
 class RegPressure {
  public:
